Updated DataProvider in order to allow adding tags when creating files and making links. Along with the DataProvide.pm, the python implementations had to be tweaked as well...
First modification impacts the “makeFiles” subroutine.
It now allows us to define the “Files” attribute as a hash as well (on top of the normal scalar definition)
Files => {
        tag1    =>      2,
        tag2    =>      4,
},

This allows us to specify appending tags at the end of LFN names.
In this example, 6 files will be generated, 2 of them in the form of xxxxxxxx-yyyyyyyy-tag1.root and 6 of them in the form of zzzzzzzzz-yyyyyyyy-tag2.root
This will be used when creating TFC rules on the receiving side and separating downloaded files from one block onto several controllers. This ensures we have more FDTWriters started (one $

The second modification impacts the “makeLinks” subroutine.

The forms of 'SrcFiles' which are now allowed in the .conf files are either:
Array form:
       SrcFiles => [
           {
               Name => '/data/Tests/links/file1.data',
               Checksum => 'cksum:1948514810',
           },
           {
               Name => '/data/Tests/links/file2.data',
               Checksum => 'cksum:1948514810',
           }
       ],

Hash form:
        SrcFiles => {
            tag1        => {
                Name => '/data/Tests/links/file1.data',
                Checksum =>  'cksum:2879015793',
            },
            tag2        => {
                Name => '/data/Tests/links/file2.data',
                Size => 16106127360,
                Checksum => 'cksum:2879015793',
            },
            tag3        => {
                Name => '/data/Tests/links/file3.data',
            },
        },

More details can be found in the example configurations